// packages/hardhat-ignition/packages/core/src/types/module.ts
export enum FutureType {
  NAMED_ARTIFACT_CONTRACT_DEPLOYMENT = "NAMED_ARTIFACT_CONTRACT_DEPLOYMENT",
  CONTRACT_DEPLOYMENT = "CONTRACT_DEPLOYMENT",
  NAMED_ARTIFACT_LIBRARY_DEPLOYMENT = "NAMED_ARTIFACT_LIBRARY_DEPLOYMENT",
  LIBRARY_DEPLOYMENT = "LIBRARY_DEPLOYMENT",
  CONTRACT_CALL = "CONTRACT_CALL",
  STATIC_CALL = "STATIC_CALL",
  ENCODE_FUNCTION_CALL = "ENCODE_FUNCTION_CALL",
  NAMED_ARTIFACT_CONTRACT_AT = "NAMED_ARTIFACT_CONTRACT_AT",
  CONTRACT_AT = "CONTRACT_AT",
  READ_EVENT_ARGUMENT = "READ_EVENT_ARGUMENT",
  SEND_DATA = "SEND_DATA",
}